address comments, consolidate transactions into one file
