--- crates/astria-proto/src/native/sequencer/v1alpha1.rs
@@ -1,19 +1,0 @@
-pub mod action;
-pub mod asset;
-pub mod mint_action;
-pub mod sequence_action;
-pub mod signed_transaction;
-pub mod sudo_address_change_action;
-pub mod transfer_action;
-pub mod types;
-pub mod unsigned_transaction;
-
-pub use action::*;
-pub use asset::*;
-pub use mint_action::*;
-pub use sequence_action::*;
-pub use signed_transaction::*;
-pub use sudo_address_change_action::*;
-pub use transfer_action::*;
-pub use types::*;
-pub use unsigned_transaction::*;
